Fixed solarized (on Linux)
Configuring tmux/solarised/vim to behave is a problem.
There is a useful discussion on the solarized github page
under issue 159. I tried to set the environment variables in
a number of different ways, but was unable to get TERM
to be anything other than "screen" under tmux. Setting
t_Co to 16 in the .vimrc fixed things.